🚚 Create app folder & move logger in it's own folder
